Refactoring: OCRScanDisplay extracted const data into separate file
