+++ tdd_intro/homework/03_bank_ocr/03_bank_ocr.pro
@@ -7,3 +7,7 @@ CONFIG -= qt
 
 SOURCES += \
     test.cpp
+
+HEADERS += \
+    constdata.h \
+    display.h
