+++ tdd_intro/homework/03_bank_ocr/constdata.h
@@ -1,0 +1,98 @@
+#pragma once
+#include "display.h"
+
+const Digit s_digit0 = { " _ ",
+                         "| |",
+                         "|_|"
+                       };
+const Digit s_digit1 = { "   ",
+                         "  |",
+                         "  |"
+                       };
+const Digit s_digit2 = { " _ ",
+                         " _|",
+                         "|_ "
+                       };
+const Digit s_digit3 = { " _ ",
+                         " _|",
+                         " _|"
+                       };
+const Digit s_digit4 = { "   ",
+                         "|_|",
+                         "  |"
+                       };
+const Digit s_digit5 = { " _ ",
+                         "|_ ",
+                         " _|"
+                       };
+const Digit s_digit6 = { " _ ",
+                         "|_ ",
+                         "|_|"
+                       };
+const Digit s_digit7 = { " _ ",
+                         "  |",
+                         "  |"
+                       };
+const Digit s_digit8 = { " _ ",
+                         "|_|",
+                         "|_|"
+                       };
+const Digit s_digit9 = { " _ ",
+                         "|_|",
+                         " _|"
+                       };
+
+const Display s_displayAll0 = { " _  _  _  _  _  _  _  _  _ ",
+                                "| || || || || || || || || |",
+                                "|_||_||_||_||_||_||_||_||_|"
+};
+
+const Display s_displayAll1 = { "                           ",
+                                "  |  |  |  |  |  |  |  |  |",
+                                "  |  |  |  |  |  |  |  |  |"
+};
+
+const Display s_displayAll2 = {  " _  _  _  _  _  _  _  _  _ ",
+                                 " _| _| _| _| _| _| _| _| _|",
+                                 "|_ |_ |_ |_ |_ |_ |_ |_ |_ "
+};
+
+const Display s_displayAll3 = { " _  _  _  _  _  _  _  _  _ ",
+                                " _| _| _| _| _| _| _| _| _|",
+                                " _| _| _| _| _| _| _| _| _|"
+};
+
+const Display s_displayAll4 = { "                           ",
+                                "|_||_||_||_||_||_||_||_||_|",
+                                "  |  |  |  |  |  |  |  |  |"
+};
+
+const Display s_displayAll5 = { " _  _  _  _  _  _  _  _  _ ",
+                                "|_ |_ |_ |_ |_ |_ |_ |_ |_ ",
+                                " _| _| _| _| _| _| _| _| _|"
+};
+
+const Display s_displayAll6 = { " _  _  _  _  _  _  _  _  _ ",
+                                "|_ |_ |_ |_ |_ |_ |_ |_ |_ ",
+                                "|_||_||_||_||_||_||_||_||_|"
+};
+
+const Display s_displayAll7 = { " _  _  _  _  _  _  _  _  _ ",
+                                "  |  |  |  |  |  |  |  |  |",
+                                "  |  |  |  |  |  |  |  |  |"
+};
+
+const Display s_displayAll8 = { " _  _  _  _  _  _  _  _  _ ",
+                                "|_||_||_||_||_||_||_||_||_|",
+                                "|_||_||_||_||_||_||_||_||_|"
+};
+
+const Display s_displayAll9 = { " _  _  _  _  _  _  _  _  _ ",
+                                "|_||_||_||_||_||_||_||_||_|",
+                                " _| _| _| _| _| _| _| _| _|"
+};
+
+const Display s_display123456789 = { "    _  _     _  _  _  _  _ ",
+                                     "  | _| _||_||_ |_   ||_||_|",
+                                     "  ||_  _|  | _||_|  ||_| _|"
+};
