+++ tdd_intro/homework/03_bank_ocr/display.h
@@ -1,0 +1,15 @@
+#pragma once
+#include <string>
+
+const unsigned short g_digitLen = 3;
+const unsigned short g_linesInDigit = 3;
+struct Digit
+{
+    std::string lines[g_linesInDigit];
+};
+
+const unsigned short g_digitsOnDisplay = 9;
+struct Display
+{
+    std::string lines[g_linesInDigit];
+};
